Добавлен Service для получение фото
